+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased](https://github.com/inspirum/balikobot-php/compare/v4.2.0...master)
+### Added
+- Added support for new options *(from documentation **v1.896** (2020-11-18))*
+    - Add **PACKAGE** request filtered by **carrier_id**
 
 
 ## [v4.2.0 (2020-11-10)](https://github.com/inspirum/balikobot-php/compare/v4.1.0...v4.2.0)